requirements.txt
django
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
joblib>=1.2.0
xgboost>=1.7.0
boto3>=1.34.0
python-dotenv>=1.0.0
